Rework ndpi_main and other patches to remove risk and category handling
The iptables modules doesn't expose the results so there is no need to incur the overhead.  Also, the kernel freeze we've been seeing was from a strcmp() passed a NULL in ndpi_check_dga_name() which was used for risk.  Now that function is a stub.  If the strcmp() isn't there, it can't fail.